style:refactoring struct
