--- app/favorite/internal/repository/db/dao/db_init.go
@@ -1,72 +1,0 @@
-package dao
-
-import (
-	"context"
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/gin-gonic/gin"
-	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
-	"gorm.io/gorm/logger"
-	"gorm.io/gorm/schema"
-
-	"github.com/CocaineCong/tangseng/config"
-	log "github.com/CocaineCong/tangseng/pkg/logger"
-)
-
-var _db *gorm.DB
-
-func InitDB() {
-	mConfig := config.Conf.MySQL
-	host := mConfig.Host
-	port := mConfig.Port
-	database := mConfig.Database
-	username := mConfig.UserName
-	password := mConfig.Password
-	charset := mConfig.Charset
-	dsn := strings.Join([]string{username, ":", password, "@tcp(", host, ":", port, ")/", database, "?charset=" + charset + "&parseTime=true"}, "")
-	err := Database(dsn)
-	if err != nil {
-		fmt.Println(err)
-		log.LogrusObj.Error(err)
-	}
-}
-
-func Database(connString string) error {
-	var ormLogger logger.Interface
-	if gin.Mode() == "debug" {
-		ormLogger = logger.Default.LogMode(logger.Info)
-	} else {
-		ormLogger = logger.Default
-	}
-	db, err := gorm.Open(mysql.New(mysql.Config{
-		DSN:                       connString, // DSN data source name
-		DefaultStringSize:         256,        // string 类型字段的默认长度
-		DisableDatetimePrecision:  true,       // 禁用 datetime 精度，MySQL 5.6 之前的数据库不支持
-		DontSupportRenameIndex:    true,       // 重命名索引时采用删除并新建的方式，MySQL 5.7 之前的数据库和 MariaDB 不支持重命名索引
-		DontSupportRenameColumn:   true,       // 用 `change` 重命名列，MySQL 8 之前的数据库和 MariaDB 不支持重命名列
-		SkipInitializeWithVersion: false,      // 根据版本自动配置
-	}), &gorm.Config{
-		Logger: ormLogger,
-		NamingStrategy: schema.NamingStrategy{
-			SingularTable: true,
-		},
-	})
-	if err != nil {
-		panic(err)
-	}
-	sqlDB, _ := db.DB()
-	sqlDB.SetMaxIdleConns(20)  // 设置连接池，空闲
-	sqlDB.SetMaxOpenConns(100) // 打开
-	sqlDB.SetConnMaxLifetime(time.Second * 30)
-	_db = db
-	migration()
-	return err
-}
-
-func NewDBClient(ctx context.Context) *gorm.DB {
-	db := _db
-	return db.WithContext(ctx)
-}
