+++ app/favorite/internal/service/favorite.go
@@ -5,9 +5,9 @@ import (
 	"sync"
 
 	"github.com/CocaineCong/tangseng/app/favorite/internal/repository/db/dao"
-	"github.com/CocaineCong/tangseng/app/favorite/internal/repository/db/model"
-	e2 "github.com/CocaineCong/tangseng/consts/e"
+	"github.com/CocaineCong/tangseng/consts/e"
 	pb "github.com/CocaineCong/tangseng/idl/pb/favorite"
+	"github.com/CocaineCong/tangseng/repository/mysql/model"
 )
 
 var FavoriteSrvIns *FavoriteSrv
@@ -25,23 +25,23 @@ func GetFavoriteSrv() *FavoriteSrv {
 }
 func (s *FavoriteSrv) FavoriteCreate(ctx context.Context, req *pb.FavoriteCreateReq) (resp *pb.FavoriteCommonResponse, err error) {
 	resp = new(pb.FavoriteCommonResponse)
-	resp.Code = e2.SUCCESS
+	resp.Code = e.SUCCESS
 	err = dao.NewFavoriteDao(ctx).CreateFavorite(req)
 	if err != nil {
 		resp.Error = err.Error()
 		return
 	}
 
-	resp.Msg = e2.GetMsg(int(resp.Code))
+	resp.Msg = e.GetMsg(int(resp.Code))
 	return
 }
 
 func (s *FavoriteSrv) FavoriteList(ctx context.Context, req *pb.FavoriteListReq) (resp *pb.FavoriteListResponse, err error) {
 	resp = new(pb.FavoriteListResponse)
 	f, err := dao.NewFavoriteDao(ctx).ListFavorite(req)
-	resp.Code = e2.SUCCESS
+	resp.Code = e.SUCCESS
 	if err != nil {
-		resp.Code = e2.ERROR
+		resp.Code = e.ERROR
 		return
 	}
 	for i := range f {
@@ -56,65 +56,65 @@ func (s *FavoriteSrv) FavoriteList(ctx context.Context, req *pb.FavoriteListReq)
 
 func (s *FavoriteSrv) FavoriteUpdate(ctx context.Context, req *pb.FavoriteUpdateReq) (resp *pb.FavoriteCommonResponse, err error) {
 	resp = new(pb.FavoriteCommonResponse)
-	resp.Code = e2.SUCCESS
+	resp.Code = e.SUCCESS
 	err = dao.NewFavoriteDao(ctx).UpdateFavorite(req)
 	if err != nil {
-		resp.Code = e2.ERROR
+		resp.Code = e.ERROR
 		resp.Error = err.Error()
 		return
 	}
 
-	resp.Msg = e2.GetMsg(int(resp.Code))
+	resp.Msg = e.GetMsg(int(resp.Code))
 	return resp, nil
 }
 
 func (s *FavoriteSrv) FavoriteDelete(ctx context.Context, req *pb.FavoriteDeleteReq) (resp *pb.FavoriteCommonResponse, err error) {
 	resp = new(pb.FavoriteCommonResponse)
-	resp.Code = e2.SUCCESS
+	resp.Code = e.SUCCESS
 	err = dao.NewFavoriteDao(ctx).DeleteFavorite(req)
 	if err != nil {
-		resp.Code = e2.ERROR
+		resp.Code = e.ERROR
 		resp.Error = err.Error()
 		return
 	}
 
-	resp.Msg = e2.GetMsg(int(resp.Code))
+	resp.Msg = e.GetMsg(int(resp.Code))
 	return
 }
 
 func (s *FavoriteSrv) FavoriteDetailCreate(ctx context.Context, req *pb.FavoriteDetailCreateReq) (resp *pb.FavoriteCommonResponse, err error) {
 	resp = new(pb.FavoriteCommonResponse)
-	resp.Code = e2.SUCCESS
+	resp.Code = e.SUCCESS
 	err = dao.NewFavoriteDetailDao(ctx).CreateFavoriteDetail(req)
 	if err != nil {
-		resp.Code = e2.ERROR
+		resp.Code = e.ERROR
 		resp.Error = err.Error()
 		return
 	}
-	resp.Msg = e2.GetMsg(int(resp.Code))
+	resp.Msg = e.GetMsg(int(resp.Code))
 	return
 }
 
 func (s *FavoriteSrv) FavoriteDetailDelete(ctx context.Context, req *pb.FavoriteDetailDeleteReq) (resp *pb.FavoriteCommonResponse, err error) {
 	resp = new(pb.FavoriteCommonResponse)
-	resp.Code = e2.SUCCESS
+	resp.Code = e.SUCCESS
 	err = dao.NewFavoriteDetailDao(ctx).DeleteFavoriteDetail(req)
 	if err != nil {
-		resp.Code = e2.ERROR
+		resp.Code = e.ERROR
 		resp.Error = err.Error()
 		return
 	}
 
-	resp.Msg = e2.GetMsg(int(resp.Code))
+	resp.Msg = e.GetMsg(int(resp.Code))
 	return
 }
 
 func (s *FavoriteSrv) FavoriteDetailList(ctx context.Context, req *pb.FavoriteDetailListReq) (resp *pb.FavoriteDetailListResponse, err error) {
 	resp = new(pb.FavoriteDetailListResponse)
-	resp.Code = e2.SUCCESS
+	resp.Code = e.SUCCESS
 	fdResp, err := dao.NewFavoriteDetailDao(ctx).ListFavoriteDetail(req)
 	if err != nil {
-		resp.Code = e2.ERROR
+		resp.Code = e.ERROR
 		return
 	}
