+++ app/search_engine/cmd/main.go
@@ -10,8 +10,7 @@ import (
 
 	"github.com/CocaineCong/tangseng/app/gateway/rpc"
 	"github.com/CocaineCong/tangseng/app/search_engine/analyzer"
-	"github.com/CocaineCong/tangseng/app/search_engine/repository/db/dao"
-	"github.com/CocaineCong/tangseng/app/search_engine/repository/starrocks/bi_dao"
+	"github.com/CocaineCong/tangseng/app/search_engine/repository/starrock/bi_dao"
 	"github.com/CocaineCong/tangseng/app/search_engine/service"
 	"github.com/CocaineCong/tangseng/config"
 	pb "github.com/CocaineCong/tangseng/idl/pb/search_engine"
@@ -22,7 +21,6 @@ import (
 func main() {
 	loading.Loading()
 	rpc.Init()
-	dao.InitDB()
 	bi_dao.InitDB()
 	analyzer.InitSeg()
