+++ app/search_engine/repository/starrock/bi_dao/db_init.go
@@ -1,4 +1,4 @@
-package dao
+package bi_dao
 
 import (
 	"context"
@@ -42,12 +42,7 @@ func Database(connString string) error {
 		ormLogger = logger.Default
 	}
 	db, err := gorm.Open(mysql.New(mysql.Config{
-		DSN:                       connString, // DSN data source name
-		DefaultStringSize:         256,        // string 类型字段的默认长度
-		DisableDatetimePrecision:  true,       // 禁用 datetime 精度，MySQL 5.6 之前的数据库不支持
-		DontSupportRenameIndex:    true,       // 重命名索引时采用删除并新建的方式，MySQL 5.7 之前的数据库和 MariaDB 不支持重命名索引
-		DontSupportRenameColumn:   true,       // 用 `change` 重命名列，MySQL 8 之前的数据库和 MariaDB 不支持重命名列
-		SkipInitializeWithVersion: false,      // 根据版本自动配置
+		DSN: connString, // DSN data source name
 	}), &gorm.Config{
 		Logger: ormLogger,
 		NamingStrategy: schema.NamingStrategy{
@@ -62,7 +57,6 @@ func Database(connString string) error {
 	sqlDB.SetMaxOpenConns(100)
 	sqlDB.SetConnMaxLifetime(time.Second * 30)
 	_db = db
-	migration()
 	return err
 }
