+++ app/search_engine/repository/starrock/starrocks_test.go
@@ -1,4 +1,4 @@
-package starrocks
+package starrock
 
 import (
 	"context"
