+++ docker-compose.yaml
@@ -122,16 +122,16 @@ services:
 #      - search_engine
 
   starrocks:
-    image: starrocks.docker.scarf.sh/starrocks/allin1-ubuntu
+    image: starrock.docker.scarf.sh/starrock/allin1-ubuntu
     restart: on-failure
-    container_name: "starrocks"
+    container_name: "starrock"
     ports:
       - "8030:8030"
       - "9030:9030"
       - "8040:8040"
     volumes:
-      - ./starrocks/fe/meta:/opt/apache-doris/fe/meta
-      - ./starrocks/fe/log:/data/deploy/starrocks/fe/log
-      - ./starrocks/be/storage:/data/deploy/starrocks/be/storage
-      - ./starrocks/be/log:/data/deploy/starrocks/be/log
+      - ./starrock/fe/meta:/opt/apache-doris/fe/meta
+      - ./starrock/fe/log:/data/deploy/starrock/fe/log
+      - ./starrock/be/storage:/data/deploy/starrock/be/storage
+      - ./starrock/be/log:/data/deploy/starrock/be/log
       - ./initdb.d:/docker-entrypoint-initdb.d/
