+++ loading/loading.go
@@ -3,6 +3,7 @@ package loading
 import (
 	"github.com/CocaineCong/tangseng/config"
 	log "github.com/CocaineCong/tangseng/pkg/logger"
+	"github.com/CocaineCong/tangseng/repository/mysql/db"
 )
 
 // Loading 全局loading
@@ -10,4 +11,5 @@ func Loading() {
 	// es.InitEs()
 	config.InitConfig()
 	log.InitLog()
+	db.InitDB()
 }
