+++ repository/mysql/db/migration.go
@@ -1,16 +1,18 @@
-package dao
+package db
 
 import (
 	"os"
 
-	"github.com/CocaineCong/tangseng/app/favorite/internal/repository/db/model"
 	log "github.com/CocaineCong/tangseng/pkg/logger"
+	"github.com/CocaineCong/tangseng/repository/mysql/model"
 )
 
 func migration() {
 	// 自动迁移模式
 	err := _db.Set("gorm:table_options", "charset=utf8mb4").
 		AutoMigrate(
+			&model.User{},
+			&model.InputData{},
 			&model.Favorite{},
 			&model.FavoriteDetail{},
 		)
