+++ repository/mysql/model/favorite.go
