+++ repository/mysql/model/favorite_detail.go
